fcntl return value an parameter fixup
fcntl returns int and not long for F_GETFL. F_SETFL expects in as well
as a parameter.

+++ src/rofl/common/crofbase.cc
@@ -280,7 +280,7 @@ crofbase::listen(
 	}
 
 	/* make socket non-blocking */
-	long flags;
+	int flags;
 	if ((flags = ::fcntl(sd, F_GETFL)) < 0) {
 		throw eSysCall("fnctl() F_GETFL");
 	}
